requirements.txt
arrow==1.3.0
asgiref==3.9.1
binaryornot==0.4.4
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
cryptography==45.0.6
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.2.23
django-allauth==0.57.2
django-browser-reload==1.18.0
django-summernote==0.8.20.0
django-tailwind==4.2.0
gunicorn==20.1.0
honcho==2.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.3.1
pillow==11.3.0
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250708
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.9.0
django-crispy-forms==2.0
crispy-tailwind==0.5.0
